Bug 1202259: Use correct filename for pofile sources.
The source path for pofiles is a special case where the source files are
pot files instead of po files.
