+++ pontoon/administration/tests/fake-checkout/templates/main.lang
