+++ pontoon/base/vcs_models.py
@@ -5,10 +5,39 @@ import logging
 import os.path
 from itertools import chain
 
-from pontoon.base import SyncError
+from django.utils.functional import cached_property
+
+from pontoon.base import MOZILLA_REPOS
+from pontoon.base.models import Resource
+from pontoon.base.utils import first
 
 
 log = logging.getLogger(__name__)
+
+
+def is_resource(filename):
+    """
+    Return True if the filename's extension is a supported Resource
+    format.
+    """
+    filename, extension = os.path.splitext(filename)
+    if extension and extension[1:] in Resource.ALLOWED_EXTENSIONS:
+        return True
+    else:
+        return False
+
+
+def directory_contains_resources(directory_path):
+    """
+    Return True if the given directory contains at least one
+    supported resource file (checked via file extension), or False
+    otherwise.
+    """
+    for root, dirnames, filenames in os.walk(directory_path):
+        # first() avoids checking past the first matching resouce.
+        if first(filenames, is_resource) is not None:
+            return True
+    return False
 
 
 class VCSProject(object):
@@ -23,15 +52,89 @@ class VCSProject(object):
         """
         self.db_project = db_project
 
-        self.resources = {}
-        for path in db_project.relative_resource_paths():
-            self.resources[path] = VCSResource(self, path)
+    @cached_property
+    def resources(self):
+        """
+        Lazy-loaded mapping of relative paths -> VCSResources.
+
+        Waiting until first access both avoids unnecessary file reads
+        and allows tests that don't need to touch the resources to run
+        with less mocking.
+        """
+        return {path: VCSResource(self, path) for path in self.relative_resource_paths()}
 
     @property
     def entities(self):
         return chain.from_iterable(
             resource.entities.values() for resource in self.resources.values()
         )
+
+    @property
+    def checkout_path(self):
+        return self.db_project.checkout_path
+
+    def source_directory_path(self):
+        """Path to the directory where source strings are stored."""
+        for root, dirnames, filenames in os.walk(self.checkout_path):
+            for dirname in dirnames:
+                if dirname in ('templates', 'en-US', 'en'):
+                    # Ensure the matched directory contains resources.
+                    directory_path = os.path.join(root, dirname)
+                    if directory_contains_resources(directory_path):
+                        return directory_path
+
+        raise Exception('No source directory found for project {0}'
+                        .format(self.db_project.slug))
+
+    def locale_directory_path(self, locale_code=None):
+        """
+        Path to the directory where strings for the given locale are
+        stored.
+
+        If locale_code is None, return the path to the directory where
+        source strings are stored.
+        """
+        path = self.checkout_path
+        locale_code = locale_code or self.source_directory_name()
+        if locale_code is not None:
+            for root, dirnames, filenames in os.walk(path):
+                if locale_code in dirnames:
+                    return os.path.join(root, locale_code)
+
+                locale_variant = locale_code.replace('-', '_')
+                if locale_variant in dirnames:
+                    return os.path.join(root, locale_variant)
+
+        raise Exception('Directory for locale `{0}` not found'.format(
+                        locale_code or 'source'))
+
+    def relative_resource_paths(self):
+        """
+        List of paths relative to the locale directories returned by
+        self.locale_directory_path() for each resource in this project.
+        """
+        path = self.source_directory_path()
+        for absolute_path in self.resources_for_path(path):
+            # .pot files in the source directory need to be renamed to
+            # .po files for the locale directories.
+            if absolute_path.endswith('.pot'):
+                absolute_path = absolute_path[:-1]
+
+            yield os.path.relpath(absolute_path, path)
+
+    def resources_for_path(self, path):
+        """
+        List of paths for all supported resources found within the given
+        path.
+        """
+        for root, dirnames, filenames in os.walk(path):
+            # Ignore certain files in Mozilla repositories.
+            if self.db_project.repository_url in MOZILLA_REPOS:
+                filenames = [f for f in filenames if f.endswith('region.properties')]
+
+            for filename in filenames:
+                if is_resource(filename):
+                    yield os.path.join(root, filename)
 
 
 class VCSResource(object):
@@ -49,10 +152,13 @@ class VCSResource(object):
         self.files = {}
         self.entities = {}
 
-        db_project = vcs_project.db_project
-
         # Create entities using resources from the source directory,
-        source_resource_path = os.path.join(db_project.source_directory_path(), self.path)
+        source_resource_path = os.path.join(vcs_project.source_directory_path(), self.path)
+
+        # Special case: Source files for pofiles are actually .pot.
+        if source_resource_path.endswith('po'):
+            source_resource_path += 't'
+
         source_resource_file = formats.parse(source_resource_path)
         for index, translation in enumerate(source_resource_file.translations):
             vcs_entity = VCSEntity(
@@ -67,9 +173,9 @@ class VCSResource(object):
             self.entities[vcs_entity.key] = vcs_entity
 
         # Fill in translations from the locale resources.
-        for locale in db_project.locales.all():
+        for locale in vcs_project.db_project.locales.all():
             resource_path = os.path.join(
-                db_project.locale_directory_path(locale.code),
+                vcs_project.locale_directory_path(locale.code),
                 self.path
             )
             try:
